omx: CB1 AmlogicAudioFFMPEGDecoder out pts value is wrong [1/1]
PD#SWPL-79667

Problem:
the first AmlogicAudioFFMPEGDecoder out pts value is wrong, then the
player renderer uses it,it cause audio jump and avsync have some problem.
The root cause is the variable mNumFramesOutput is not initialized, so
when using it first time it's value will be a random value.

Solution:
initialize mNumFramesOutput  in constructed function.

Verify:
AndroidR

Change-Id: I26a489a60aebe25ccd5cd12dfc4eed00ef78e4bc
Signed-off-by: Qiang Guo <qiang.guo@amlogic.com>