screencontrol: fix screencap take too long. [1/1]
PD#SWPL-156815

Problem:
When the screenshot action occurs in the process
of recording the screen, it is necessary to use
the software algorithm to convert the original
resolution yuv data into the RGB data needed for
the screenshot, so it will lead to a long time.

Solution:
Raw data formats are converted using G2ED hardware.

Verify:
T982_AR301

Change-Id: Ie8c44761c5e7ded22a89bef3f6f97664443263cc
Signed-off-by: zhewen.luo <zhewen.luo@amlogic.com>